requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
streamlit==1.30.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0

# AI/ML
openai==1.10.0
google-generativeai==0.3.2
chromadb==0.4.22
tiktoken==0.5.2

# Validation & Config
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
requests==2.31.0
